Calculate difference between actual elapsed time and timer interval
Use ElapsedTicks to get more precision for elapsed time from the Stopwatch
and use that to compare with the 100 ms interval for the Timer.

Testing shows that the difference between actual elapsed time and 100 ms
interval is about 9%, meaning actual elapsed time measured with Stopwatch
is 109 ms while the timer interval is 100 ms.